+++ public/js/ws.js
@@ -20,6 +20,7 @@ import { setHomeDir, refreshPanePath } from "./pane-path.js";
 import { markPaneReady } from "./pane-boot.js";
 import { setTreeRoot, setTreeEntries, setTreeStatus, treeChanged, syncTreeRoot, fileTreeReconnected } from "./file-tree.js";
 import { setFileView, viewChanged, viewerReconnected } from "./file-view.js";
+import { setFileList } from "./quick-open.js";
 
 const RECONNECT_MS = 1000;
 const PROBE_TIMEOUT_MS = 3000;
@@ -177,6 +178,9 @@ export function connect() {
         break;
       case "viewChanged":
         viewChanged(msg);
+        break;
+      case "fileList":
+        setFileList(msg);
         break;
       case "fileContent":
         setMdContent(msg);